app/requirements.txt
asyncpg==0.29.0
fastapi==0.136.0
pydantic==2.13.2
python-dotenv==1.2.2
sqlalchemy==2.0.49
uvicorn==0.44.0
